+++ packages/hooks/src/index.ts
@@ -1,1 +1,5 @@
 export * from './useApi';
+export * from './createContext';
+export * from './usePortal';
+export * from './useIntersection';
+export * from './useResize';