+++ packages/libraries/web/index.ts
@@ -6,5 +6,7 @@ export * from '@elonwu/web-empty';
 export * from '@elonwu/web-form';
 export * from '@elonwu/web-graph';
 export * from '@elonwu/web-icon';
+export * from '@elonwu/web-menu';
 export * from '@elonwu/web-tag';
 export * from '@elonwu/web-text';
+export * from '@elonwu/web-portal';