+++ packages/web/slider/rollup.config.js
@@ -1,10 +1,9 @@
-import svg from 'rollup-plugin-svg';
 import pkg from './package.json';
-import svg from 'rollup-plugin-svg';
 import { override } from '../../../rollup.config';
 
 const config = override(pkg, ({ tsConfig, dtsConfig }) => [
-  Object.assign({}, tsConfig, { plugins: [svg(), ...tsConfig.plugins] }),
+  tsConfig,
   dtsConfig,
 ]);
+
 export default config;