+++ packages/libraries/web/CHANGELOG.md
@@ -3,6 +3,60 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.7](https://github.com/ElonWu/elonwu_ui/compare/@elonwu/web@0.4.6...@elonwu/web@0.4.7) (2021-10-13)
+
+**Note:** Version bump only for package @elonwu/web
+
+
+
+
+
+## [0.4.6](https://github.com/ElonWu/elonwu_ui/compare/@elonwu/web@0.4.5...@elonwu/web@0.4.6) (2021-10-13)
+
+**Note:** Version bump only for package @elonwu/web
+
+
+
+
+
+## [0.4.5](https://github.com/ElonWu/elonwu_ui/compare/@elonwu/web@0.4.4...@elonwu/web@0.4.5) (2021-10-12)
+
+**Note:** Version bump only for package @elonwu/web
+
+
+
+
+
+## [0.4.4](https://github.com/ElonWu/elonwu_ui/compare/@elonwu/web@0.4.3...@elonwu/web@0.4.4) (2021-10-12)
+
+
+### Bug Fixes
+
+* web 导出 portal ([bfe24e5](https://github.com/ElonWu/elonwu_ui/commit/bfe24e5ab59c72c76653852793310a7001d4acab))
+
+
+
+
+
+## [0.4.3](https://github.com/ElonWu/elonwu_ui/compare/@elonwu/web@0.4.2...@elonwu/web@0.4.3) (2021-10-12)
+
+**Note:** Version bump only for package @elonwu/web
+
+
+
+
+
+## [0.4.2](https://github.com/ElonWu/elonwu_ui/compare/@elonwu/web@0.4.1...@elonwu/web@0.4.2) (2021-10-09)
+
+
+### Bug Fixes
+
+* 配置 menu 打包 ([63b367d](https://github.com/ElonWu/elonwu_ui/commit/63b367db606501f8bb42906b5d8686e4022554bf))
+
+
+
+
+
 ## [0.4.1](https://github.com/ElonWu/elonwu_ui/compare/@elonwu/web@0.4.0...@elonwu/web@0.4.1) (2021-10-09)
 
 **Note:** Version bump only for package @elonwu/web